Fix critical bugs in Marklin CAN protocol implementation
This commit fixes several critical bugs in the Marklin CAN (MCAN) protocol
implementation that were causing incorrect message parsing and test failures.
All fixes have been verified against the CS2 CAN Protocol 2.0 specification.

Critical Protocol Fixes:
1. Fix priority extraction in MarklinCanCodec.getPriority()
   - Changed from >> 6 (2-bit) to >> 4 (4-bit extraction)
   - Priority field occupies bits 7-4 of byte 0 per CS2 spec
   - Aligns with encoder implementation at line 170

2. Fix DLC (Data Length Code) encoding in MessageBuilder.build()
   - Removed incorrect +4 offset: DLC represents data bytes only (0-8)
   - DLC does NOT include the 4 address bytes per CS2 spec
   - Line 180: Changed from (dataLength + 4) to dataLength

3. Fix DLC decoding in MarklinCanCodec.decode()
   - Removed incorrect -4 offset from DLC parsing
   - Lines 222-226: Changed from (rawMessage[4] - 4) to rawMessage[4]
   - Now correctly interprets DLC as actual data byte count

4. Fix ArrayIndexOutOfBoundsException in MarklinMessage.getCanBoot()
   - Removed incorrect setDataLength(5) call after setData(0x11)
   - setData() already sets dataLength=1 automatically
   - Prevents writing beyond 13-byte message boundary

5. Clean up MarklinMessage.getCanBootloaderMode()
   - Removed redundant setDataLength(0) call (already default)
   - Added clarifying comment about DLC=0 for bootloader mode

Convenience Methods Added:
- Add decode(MarklinReply) convenience method to MarklinCanCodec
- Add decode(MarklinMessage) convenience method to MarklinCanCodec
- Eliminates repetitive 13-element byte array copying in client code
- Used in MarklinThrottle and MarklinTurnout (10 lines eliminated)

Test Fixes:
- MarklinTurnoutTest: Add explicit priority setup in test messages
- MarklinMessageTest: Update DLC expectation from 5 to 1

Code Quality Improvements (from ECJ static analysis):
- Remove unnecessary cast in MarklinCanCodec.java:232
- Remove unnecessary cast in MarklinDeviceIdentifier.java:189
- Remove unused import in CdbSerialToTcpBridge.java:4

Verification:
- All Marklin unit tests pass: MarklinCanCodecTest (32 tests),
  MarklinTurnoutTest, MarklinMessageTest
- Full headless test suite: BUILD SUCCESSFUL (31m 43s)
- ECJ compiler warnings: 3 issues found and resolved
- Changes verified against CS2 CAN Protocol 2.0 specification

Files Modified:
- MarklinCanCodec.java: Priority, DLC fixes, convenience methods
- MarklinMessage.java: Boot message fixes
- MarklinDeviceIdentifier.java: Code quality fix
- CdbSerialToTcpBridge.java: Code quality fix
- MarklinTurnoutTest.java: Test fixes
- MarklinMessageTest.java: Test expectation updates

+++ java/src/jmri/jmrix/marklin/MarklinDeviceIdentifier.java
@@ -186,7 +186,7 @@ public class MarklinDeviceIdentifier {
         long uid = ((long) decoded.getDataByte(0) << 24)
                  | ((long) decoded.getDataByte(1) << 16)
                  | ((long) decoded.getDataByte(2) << 8)
-                 | ((long) decoded.getDataByte(3));
+                 | decoded.getDataByte(3);
 
         // Extract software version (bytes 4-5)
         int softwareVersion = (decoded.getDataByte(4) << 8) | decoded.getDataByte(5);
